% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{OpenWeather-class}
\alias{OpenWeather}
\alias{OpenWeather-class}
\title{OpenWeather-class
Contains requested weather for a given list of locations}
\description{
OpenWeather-class
Contains requested weather for a given list of locations
}
\section{Slots}{

\describe{
\item{\code{cityInfo}}{List providing city information}

\item{\code{weather}}{data.frame with weather observations}

\item{\code{obsCount}}{Count of observation}
}}

